refactor methods and properties we think we will need in other fragments to a super class.
